Cleaning and texturing
Cleaned up the Renderer class a bit more, swapped back from the
halfspace triangle rasterize function to the scanline for the time being
as I have no clue how to interpolate coords for the texture mapping
using halfspace (even though it far out performs scanline (at least for
flat tris). Also added a rather buggy affine texture mapping to the
scanline draw triangle function.

+++ SoftwareRenderer/Bitmap.h
@@ -47,7 +47,7 @@ public:
 	VOID LoadFromFile(const WCHAR *path);
 	VOID Release();
 
-	BYTE *data;
+	DWORD *data;
 	RGBQUAD *palette;
 	DWORD size;
 	BITMAPFILEHEADER fileHeader;
